+++ src/main/java/gtexpert/gtwp/api/util/Mods.java
@@ -92,6 +92,7 @@ public enum Mods {
     ProjectRedIllumination(Names.PROJECT_RED_ILLUMINATION),
     Railcraft(Names.RAILCRAFT),
     RefinedStorage(Names.REFINED_STORAGE),
+    ProjectVibrantJourneys(Names.PROJECT_VIBRANT_JOURNEYS),
     Thaumcraft(Names.THAUMCRAFT),
     ThaumicEnergistics(Names.THAUMIC_ENERGISTICS),
     TheOneProbe(Names.THE_ONE_PROBE),
@@ -188,6 +189,7 @@ public enum Mods {
         public static final String PROJECT_RED_ILLUMINATION = "projectred-illumination";
         public static final String RAILCRAFT = "railcraft";
         public static final String REFINED_STORAGE = "refinedstorage";
+        public static final String PROJECT_VIBRANT_JOURNEYS = "pvj";
         public static final String THAUMCRAFT = "thaumcraft";
         public static final String THAUMIC_ENERGISTICS = "haumicenergistics";
         public static final String THE_ONE_PROBE = "theoneprobe";