+++ src/main/java/gtexpert/gtwp/integration/tc/recipes/TCWoodRecipe.java
@@ -3,7 +3,6 @@ package gtexpert.gtwp.integration.tc.recipes;
 import java.util.Arrays;
 import java.util.List;
 
-import gregtech.api.recipes.ModHandler;
 import gregtech.loaders.WoodTypeEntry;
 import gtexpert.gtwp.api.util.Mods;
 import gtexpert.gtwp.loaders.GTWPWoodRecipeLoader;
@@ -14,26 +13,28 @@ public class TCWoodRecipe {
 
     private static List<WoodTypeEntry> getDefaultEntries() {
         if (DEFAULT_ENTRIES == null) {
-            final String mcModId = Mods.Thaumcraft.name();
+            final String mcModId = Mods.Names.THAUMCRAFT;
             return DEFAULT_ENTRIES = Arrays.asList(
                     new WoodTypeEntry.Builder(mcModId, "silverwood")
-                            .planks(Mods.Thaumcraft.getItem("plank_silverwood", 1), null)
-                            .log(Mods.Thaumcraft.getItem("log_silverwood", 1))
-                            .registerAllUnificationInfo()
+                            .log(Mods.Thaumcraft.getItem("log_silverwood", 1)).removeCharcoalRecipe()
+                            .planks(Mods.Thaumcraft.getItem("plank_silverwood", 1), "planksilverwood")
+                            .slab(Mods.Thaumcraft.getItem("slab_silverwood", 1), "slabsilverwood")
+                            .stairs(Mods.Thaumcraft.getItem("stairs_silverwood", 1), "stairssilverwood")
                             .build(),
                     new WoodTypeEntry.Builder(mcModId, "greatwood")
-                            .planks(Mods.Thaumcraft.getItem("plank_greatwood", 1), null)
-                            .log(Mods.Thaumcraft.getItem("log_greatwood", 1))
+                            .log(Mods.Thaumcraft.getItem("log_greatwood", 1)).removeCharcoalRecipe()
+                            .planks(Mods.Thaumcraft.getItem("plank_greatwood", 1), "plankgreatwood")
+                            .slab(Mods.Thaumcraft.getItem("slab_greatwood", 1), "slabgreatwood")
+                            .stairs(Mods.Thaumcraft.getItem("stairs_greatwood", 1), "stairsgreatwood")
                             .build());
         }
         return DEFAULT_ENTRIES;
     }
 
     public static void init() {
-        ModHandler.removeRecipeByName(Mods.Thaumcraft.getResource("planksilverwood"));
-        ModHandler.removeRecipeByName(Mods.Thaumcraft.getResource("plankgreatwood"));
-
         for (WoodTypeEntry entry : getDefaultEntries()) {
+            GTWPWoodRecipeLoader.removePlankRecipe(false, entry, Mods.Names.THAUMCRAFT);
+
             GTWPWoodRecipeLoader.registerWoodTypeRecipe(false, entry);
             GTWPWoodRecipeLoader.addCuttingRecipe(entry);
             GTWPWoodRecipeLoader.addSawmillRecipe(entry);