+++ src/main/java/gtexpert/gtwp/module/Modules.java
@@ -32,6 +32,7 @@ public class Modules implements IModuleContainer {
     public static final String MODULE_AVARITIA = "avaritia_integration";
     public static final String MODULE_TC = "tc_integration";
     public static final String MODULE_TE = "te_integration";
+    public static final String MODULE_PVJ = "pvj_integration";
 
     @Override
     public String getID() {