+++ CHANGELOG.md
@@ -1,2 +1,17 @@
+# v1.0.2
+## New Integration
+- Project: Vibrant Journeys Integration
+
+* * *
+
+## Fix Integration
+- Thaumcraft
+
+# v1.0.1
+## New Integration
+- Advanced Rocketry
+
+* * *
+
 # v1.0.0
 - Fast version